+++ stocks/stock.go
@@ -4,7 +4,7 @@ type Stock struct {
 	Title string
 	Description string
 	Sect Sector
-	TotalSupply float32
-	StockVal float32
-	ID uint64
+	TotalSupply float64
+	StockVal float64
+	ID uint64 // Is it necessary?
 }
